# last_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MusicAI.Orchestrator",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Agent/chat": {
      "post": {
        "tags": [
          "Agent"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Playback/oap/{userId}": {
      "get": {
        "tags": [
          "Playback"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "persona",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "conversationHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringStringValueTuple"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StringStringValueTuple": {
        "type": "object",
        "additionalProperties": false
      },
...
    }
  }
}